+++ buession-security-crypto/src/main/java/com/buession/security/crypto/passwordgenerator/HmacMd5PasswordGenerator.java
@@ -22,7 +22,7 @@
  * | Copyright @ 2013-2023 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
-package com.buession.security.mcrypt.passwordgenerator;
+package com.buession.security.crypto.passwordgenerator;
 
 import com.buession.security.crypto.HmacMD5Crypto;
 
@@ -30,11 +30,13 @@ import com.buession.security.crypto.HmacMD5Crypto;
  * HmacMD5 密码生成器
  *
  * @author Yong.Teng
- * @see com.buession.security.crypto.passwordgenerator.HmacMd5PasswordGenerator
- * @since 2.1.0
+ * @since 2.3.3
  */
 public class HmacMd5PasswordGenerator extends AbstractPasswordGenerator {
 
+	/**
+	 * 构造函数
+	 */
 	public HmacMd5PasswordGenerator() {
 		super(new HmacMD5Crypto());
 	}