+++ buession-security-web/src/main/java/com/buession/security/web/xss/factory/EscapeXssFactory.java
@@ -19,24 +19,35 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2023 Buession.com Inc.														       |
+ * | Copyright @ 2013-2024 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
-package com.buession.security.mcrypt.passwordgenerator;
+package com.buession.security.web.xss.factory;
 
-import com.buession.security.crypto.HmacMD5Crypto;
+import com.buession.security.web.xss.Options;
+import org.apache.commons.text.StringEscapeUtils;
 
 /**
- * HmacMD5 密码生成器
+ * 转义模式 XSS 处理工厂
  *
  * @author Yong.Teng
- * @see com.buession.security.crypto.passwordgenerator.HmacMd5PasswordGenerator
- * @since 2.1.0
+ * @since 2.3.3
  */
-public class HmacMd5PasswordGenerator extends AbstractPasswordGenerator {
+public class EscapeXssFactory extends AbstractXssFactory {
 
-	public HmacMd5PasswordGenerator() {
-		super(new HmacMD5Crypto());
+	/**
+	 * 构造函数
+	 *
+	 * @param options
+	 * 		XSS 处理选项
+	 */
+	public EscapeXssFactory(Options options) {
+		super(options);
+	}
+
+	@Override
+	public String handle(String str) {
+		return StringEscapeUtils.escapeHtml4(str);
 	}
 
 }