+++ buession-security-web/src/main/java/com/buession/security/web/xss/servlet/XssFilter.java
@@ -21,12 +21,13 @@
  * +------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										|
  * | Author: Yong.Teng <webmaster@buession.com> 													|
- * | Copyright @ 2013-2023 Buession.com Inc.														|
+ * | Copyright @ 2013-2024 Buession.com Inc.														|
  * +------------------------------------------------------------------------------------------------+
  */
 package com.buession.security.web.xss.servlet;
 
-import org.owasp.validator.html.Policy;
+import com.buession.core.utils.Assert;
+import com.buession.security.web.xss.Options;
 import org.springframework.lang.Nullable;
 import org.springframework.web.filter.OncePerRequestFilter;
 
@@ -44,22 +45,69 @@ import java.io.IOException;
  */
 public class XssFilter extends OncePerRequestFilter {
 
-	private Policy policy;
+	/**
+	 * XSS 处理选项
+	 */
+	private Options options = Options.Builder.getInstance().build();
 
-	public Policy getPolicy() {
-		return policy;
+	/**
+	 * 构造函数
+	 */
+	public XssFilter() {
 	}
 
-	public void setPolicy(Policy policy) {
-		this.policy = policy;
+	/**
+	 * 构造函数
+	 *
+	 * @param options
+	 *        {@link Options}
+	 *
+	 * @since 2.3.3
+	 */
+	public XssFilter(final Options options) {
+		setOptions(options);
+	}
+
+	/**
+	 * 返回 XSS 处理选项
+	 *
+	 * @return XSS 处理选项
+	 *
+	 * @since 2.3.3
+	 */
+	public Options getOptions() {
+		return options;
+	}
+
+	/**
+	 * 设置 XSS 处理选项
+	 *
+	 * @param options
+	 * 		XSS 处理选项
+	 *
+	 * @since 2.3.3
+	 */
+	public void setOptions(Options options) {
+		Assert.isNull(options, "Options cloud not be null");
+		this.options = options;
+	}
+
+	/**
+	 * 设置策略
+	 *
+	 * @param policy
+	 * 		策略
+	 *
+	 * @since 2.3.3
+	 */
+	public void setPolicy(Options.Policy policy) {
+		getOptions().setPolicy(policy);
 	}
 
 	@Override
 	protected void doFilterInternal(@Nullable HttpServletRequest request, @Nullable HttpServletResponse response,
-									FilterChain filterChain)
-			throws ServletException, IOException {
-		XssServletRequestWrapper xssServletRequestWrapper = new XssServletRequestWrapper(request, getPolicy());
-		filterChain.doFilter(xssServletRequestWrapper, response);
+									FilterChain filterChain) throws ServletException, IOException {
+		filterChain.doFilter(new XssServletRequestWrapper(request, getOptions()), response);
 	}
 
 }